update expired time jwt
