+++ src/main/java/org/example/likelion/service/auth/AuthenticationServiceImpl.java
@@ -15,11 +15,9 @@ import org.example.likelion.model.Token;
 import org.example.likelion.model.User;
 import org.example.likelion.repository.TokenRepository;
 import org.example.likelion.repository.UserRepository;
-import org.example.likelion.service.jwt.IJwtService;
 import org.example.likelion.service.jwt.JwtService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -38,7 +36,7 @@ public class AuthenticationServiceImpl implements AuthenticationService {
     PasswordEncoder encoder;
 
     @Autowired
-    IJwtService jwtService;
+    JwtService jwtService;
 
     @Autowired
     UserRepository userRepository;
