+++ src/main/java/org/example/likelion/service/jwt/JwtService.java
@@ -1,121 +1,25 @@
 package org.example.likelion.service.jwt;
 
-import io.jsonwebtoken.*;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import io.jsonwebtoken.security.SignatureException;
-import org.example.likelion.model.User;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Value;
+import io.jsonwebtoken.Claims;
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Service;
 
-import javax.crypto.SecretKey;
-import java.security.Key;
-import java.util.Date;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Function;
 
-@Service
-public class JwtService implements IJwtService {
-    @Value(value = "${baeldung.app.jwtSecret}")
-    private String secretKey;
-    private static final Logger logger = LoggerFactory.getLogger(JwtService.class);
-    @Value("${baeldung.app.jwtExpirationMs}")
-    private int jwtExpiration;
-    @Value("${baeldung.app.jwtExpirationMs}")
-    private int refreshExpiration;
+public interface JwtService {
+    String extractUsername(String token);
 
-    public String extractUsername(String token) {
-        return extractClaim(token, Claims::getSubject);
-    }
+    <T> T extractClaim(String token, Function<Claims, T> claimsResolver);
 
-    public <T> T extractClaim(String token, Function<Claims, T> claimsResolver) {
-        final Claims claims = extractAllClaims(token);
-        return claimsResolver.apply(claims);
-    }
+    String generateToken(UserDetails userDetails);
 
-    public String generateToken(UserDetails userDetails) {
-        return buildToken(new HashMap<>(), userDetails, jwtExpiration);
-    }
+    String generateToken(UserDetails userDetails, Map<String, Object> extraClaims);
 
-    public String generateToken(UserDetails userDetails, Map<String, Object> extraClaims) {
-        return buildToken(extraClaims, userDetails, jwtExpiration);
-    }
+    String generateRefreshToken(UserDetails userDetails);
 
-    public String generateRefreshToken(UserDetails userDetails) {
-        return buildToken(new HashMap<>(), userDetails, refreshExpiration);
-    }
+    String getUserNameFromJwtToken(String token);
 
-    private String buildToken(Map<String, Object> extraClaims,
-                              UserDetails userDetails,
-                              long expiration) {
-        return Jwts.builder()
-                .subject((userDetails.getUsername()))
-                .claims(extraClaims)
-                .issuedAt(new Date())
-                .expiration(new Date((new Date()).getTime() + expiration))
-                .signWith(getSigningKey())
-                .compact();
-    }
+    boolean validateJwtToken(String authToken);
 
-    private SecretKey getSigningKey() {
-        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
-        return Keys.hmacShaKeyFor(keyBytes);
-    }
-
-    public String getUserNameFromJwtToken(String token) {
-        return Jwts.parser()
-                .verifyWith(getSigningKey())
-                .build()
-                .parseSignedClaims(token)
-                .getPayload()
-                .getSubject();
-    }
-
-    public boolean validateJwtToken(String authToken) {
-        try {
-            Jwts.parser()
-                    .verifyWith(getSigningKey())
-                    .build()
-                    .parseSignedClaims(authToken);
-            return true;
-        } catch (SignatureException e) {
-            logger.error("Invalid JWT signature: {}", e.getMessage());
-        } catch (MalformedJwtException e) {
-            logger.error("Invalid JWT token: {}", e.getMessage());
-        } catch (ExpiredJwtException e) {
-            logger.error("JWT token is expired: {}", e.getMessage());
-        } catch (UnsupportedJwtException e) {
-            logger.error("JWT token is unsupported: {}", e.getMessage());
-        } catch (IllegalArgumentException e) {
-            logger.error("JWT claims string is empty: {}", e.getMessage());
-        }
-        return false;
-    }
-
-
-    public boolean isTokenValid(String token, UserDetails userDetails) {
-        final String username = extractUsername(token);
-        return (username.equals(userDetails.getUsername())) && !isTokenExpired(token);
-    }
-
-    private boolean isTokenExpired(String token) {
-        return extractExpiration(token).before(new Date());
-    }
-
-    private Date extractExpiration(String token) {
-        return extractClaim(token, Claims::getExpiration);
-    }
-
-    private Claims extractAllClaims(String token) {
-        return Jwts
-                .parser()
-                .setSigningKey(getSigningKey())
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
+    boolean isTokenValid(String token, UserDetails userDetails);
 }
