+++ src/main/resources/application-prod.yml
@@ -19,7 +19,10 @@ spring:
         format_sql: true
         use_nationalized_character_data: true
     show-sql: true
-baeldung:
+likelion:
   app:
-    jwtSecret: 404E635266556A586E3272357538782F413F4428472B4B6250645367566B5970
-    jwtExpirationMs: 8640000
+    jwt:
+      jwtSecret: 404E635266556A586E3272357538782F413F4428472B4B6250645367566B5970
+      expiration: 86400000
+      refresh-token:
+        expiration: 604800000
